pyproject.toml
[project]
name = "decision-helper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "duckduckgo-search>=7.3.2",
    "ipykernel>=6.29.5",
    "pydantic>=2.10.6",
    "pydantic-ai[logfire]>=0.0.23",
    "pydantic-graph>=0.0.23",
    "python-dotenv>=1.0.1",
    "rich>=13.9.4",
]
